@@ -9,15 +9,13 @@ import RecommendPage from "../views/RecommendPage";
 const Router = () => {
   return (
     <>
-      <BrowserRouter>
-        <Routes>
-          <Route path="/intro" element={<Intro />}></Route>
-          <Route path="/business" element={<Business />}></Route>
-          <Route path="/mypage" element={<MyPage />}></Route>
-          <Route path="/chatbot" element={<ChatBot />}></Route>
-          <Route path="/recommendpage" element={<RecommendPage />}></Route>
-        </Routes>
-      </BrowserRouter>
+      <Routes>
+        <Route path="/intro" element={<Intro />} />
+        <Route path="/business" element={<Business />} />
+        <Route path="/mypage" element={<MyPage />} />
+        <Route path="/chatbot" element={<ChatBot />} />
+        <Route path="/recommendpage" element={<RecommendPage />} />
+      </Routes>
     </>
   );
 };
